requirement.txt
six==1.15.0
websocket-client==0.57.0
